Split remaining versions model to new file
Complete the remainder of the partial split out of the versions
model into a new file for readability.
