+++ models/dataset.go
@@ -3,7 +3,7 @@ package models
 import (
 	"context"
 	//nolint:gosec // not used for secure purposes.
-	"crypto/sha1"
+
 	"encoding/json"
 	"fmt"
 	"io"
@@ -14,10 +14,14 @@ import (
 
 	"github.com/pkg/errors"
 	uuid "github.com/satori/go.uuid"
-	"go.mongodb.org/mongo-driver/bson"
 
 	errs "github.com/ONSdigital/dp-dataset-api/apierrors"
 	"github.com/ONSdigital/log.go/v2/log"
+)
+
+// List of error variables
+var (
+	ErrEditionLinksInvalid = errors.New("editions links do not exist")
 )
 
 // DatasetType defines possible dataset types
@@ -67,14 +71,6 @@ func GetDatasetType(datasetType string) (DatasetType, error) {
 		return Invalid, errs.ErrDatasetTypeInvalid
 	}
 }
-
-// List of error variables
-var (
-	ErrAssociatedVersionCollectionIDInvalid = errors.New("missing collection_id for association between version and a collection")
-	ErrPublishedVersionCollectionIDInvalid  = errors.New("unexpected collection_id in published version")
-	ErrVersionStateInvalid                  = errors.New("incorrect state, can be one of the following: edition-confirmed, associated or published")
-	ErrEditionLinksInvalid                  = errors.New("editions links do not exist")
-)
 
 // DatasetUpdate represents an evolving dataset with the current dataset and the updated dataset
 // Note: Stored as Dataset (in `dataset` Collection) in MongoDB
@@ -181,219 +177,11 @@ type Edition struct {
 	QualityDesignation QualityDesignation  `bson:"quality_designation,omitempty" json:"quality_designation,omitempty"`
 }
 
-// Define the enum type for QualityDesignation
-type QualityDesignation string
-
-// Define the possible values for the QualityDesignation enum
-const (
-	QualityDesignationExperimental QualityDesignation = "experimental"
-	QualityDesignationOfficial     QualityDesignation = "official"
-	QualityDesignationAccredited   QualityDesignation = "accredited"
-)
-
-func (qd *QualityDesignation) IsValid() bool {
-	switch *qd {
-	case QualityDesignationExperimental, QualityDesignationOfficial, QualityDesignationAccredited:
-		return true
-	default:
-		return false
-	}
-}
-
-// MarshalJSON marshals the QualityDesignation to JSON
-func (qd QualityDesignation) MarshalJSON() ([]byte, error) {
-	if !qd.IsValid() {
-		return nil, fmt.Errorf("invalid QualityDesignation: %s", qd)
-	}
-	return json.Marshal(string(qd))
-}
-
-// UnmarshalJSON unmarshals a string to QualityDesignation
-func (qd *QualityDesignation) UnmarshalJSON(data []byte) error {
-	var str string
-	if err := json.Unmarshal(data, &str); err != nil {
-		return err
-	}
-	converted := QualityDesignation(str)
-	if !converted.IsValid() {
-		return fmt.Errorf("invalid QualityDesignation: %s", str)
-	}
-	*qd = converted
-	return nil
-}
-
 // Publisher represents an object containing information of the publisher
 type Publisher struct {
 	HRef string `bson:"href,omitempty" json:"href,omitempty"`
 	Name string `bson:"name,omitempty" json:"name,omitempty"`
 	Type string `bson:"type,omitempty" json:"type,omitempty"`
-}
-
-// Hash generates a SHA-1 hash of the version struct. SHA-1 is not cryptographically safe,
-// but it has been selected for performance as we are only interested in uniqueness.
-// ETag field value is ignored when generating a hash.
-// An optional byte array can be provided to append to the hash.
-// This can be used, for example, to calculate a hash of this version and an update applied to it.
-func (v *Version) Hash(extraBytes []byte) (string, error) {
-	//nolint:gosec // sha1 not used for secure purposes
-	h := sha1.New()
-
-	// copy by value to ignore ETag without affecting v
-	v2 := *v
-	v2.ETag = ""
-
-	versionBytes, err := bson.Marshal(v2)
-	if err != nil {
-		return "", err
-	}
-
-	if _, err := h.Write(append(versionBytes, extraBytes...)); err != nil {
-		return "", err
-	}
-
-	return fmt.Sprintf("%x", h.Sum(nil)), nil
-}
-
-// AlertType defines possible types of alerts
-type AlertType string
-
-const (
-	AlertTypeAlert      AlertType = "alert"
-	AlertTypeCorrection AlertType = "correction"
-)
-
-func (at AlertType) IsValid() bool {
-	switch at {
-	case AlertTypeAlert, AlertTypeCorrection:
-		return true
-	default:
-		return false
-	}
-}
-
-func (at AlertType) String() string {
-	return string(at)
-}
-
-// Alert represents an object containing information on an alert
-type Alert struct {
-	Date        string    `bson:"date,omitempty"        json:"date,omitempty"`
-	Description string    `bson:"description,omitempty" json:"description,omitempty"`
-	Type        AlertType `bson:"type,omitempty"        json:"type,omitempty"`
-}
-
-// MarshalJSON marshals the AlertType to JSON
-func (at AlertType) MarshalJSON() ([]byte, error) {
-	if !at.IsValid() {
-		return nil, fmt.Errorf("invalid AlertType: %s", at)
-	}
-	return json.Marshal(string(at))
-}
-
-// UnmarshalJSON unmarshals a string to AlertType
-func (at *AlertType) UnmarshalJSON(data []byte) error {
-	var str string
-	if err := json.Unmarshal(data, &str); err != nil {
-		return err
-	}
-	converted := AlertType(str)
-	if !converted.IsValid() {
-		return fmt.Errorf("invalid AlertType: %s", str)
-	}
-	*at = converted
-	return nil
-}
-
-// DownloadList represents a list of objects of containing information on the downloadable files.
-// Items are in a specific order and should not be changed (xls, xlsx, csv, txt, csvw)
-type DownloadList struct {
-	XLS  *DownloadObject `bson:"xls,omitempty" json:"xls,omitempty"`
-	XLSX *DownloadObject `bson:"xlsx,omitempty" json:"xlsx,omitempty"`
-	CSV  *DownloadObject `bson:"csv,omitempty" json:"csv,omitempty"`
-	TXT  *DownloadObject `bson:"txt,omitempty" json:"txt,omitempty"`
-	CSVW *DownloadObject `bson:"csvw,omitempty" json:"csvw,omitempty"`
-}
-
-// DownloadObject represents information on the downloadable file
-type DownloadObject struct {
-	HRef    string `bson:"href,omitempty"  json:"href,omitempty"`
-	Private string `bson:"private,omitempty" json:"private,omitempty"`
-	Public  string `bson:"public,omitempty" json:"public,omitempty"`
-	// TODO size is in bytes and probably should be an int64 instead of a string this
-	// will have to change for several services (filter API, exporter services and web)
-	Size string `bson:"size,omitempty" json:"size,omitempty"`
-}
-
-// LatestChange represents an object contining
-// information on a single change between versions
-type LatestChange struct {
-	Description string `bson:"description,omitempty" json:"description,omitempty"`
-	Name        string `bson:"name,omitempty"        json:"name,omitempty"`
-	Type        string `bson:"type,omitempty"        json:"type,omitempty"`
-}
-
-// TemporalFrequency represents a frequency for a particular period of time
-type TemporalFrequency struct {
-	EndDate   string `bson:"end_date,omitempty"    json:"end_date,omitempty"`
-	Frequency string `bson:"frequency,omitempty"   json:"frequency,omitempty"`
-	StartDate string `bson:"start_date,omitempty"  json:"start_date,omitempty"`
-}
-
-// UsageNote represents a note containing extra information associated to the resource
-type UsageNote struct {
-	Note  string `bson:"note,omitempty"     json:"note,omitempty"`
-	Title string `bson:"title,omitempty"    json:"title,omitempty"`
-}
-
-// VersionLinks represents a list of specific links related to the version resource for an edition of a dataset
-type VersionLinks struct {
-	Dataset    *LinkObject `bson:"dataset,omitempty"     json:"dataset,omitempty"`
-	Dimensions *LinkObject `bson:"dimensions,omitempty"  json:"dimensions,omitempty"`
-	Edition    *LinkObject `bson:"edition,omitempty"     json:"edition,omitempty"`
-	Self       *LinkObject `bson:"self,omitempty"        json:"self,omitempty"`
-	Spatial    *LinkObject `bson:"spatial,omitempty"     json:"spatial,omitempty"`
-	Version    *LinkObject `bson:"version,omitempty"     json:"-"`
-}
-
-func (vl *VersionLinks) DeepCopy() *VersionLinks {
-	dst := &VersionLinks{}
-	if vl.Dataset != nil {
-		dst.Dataset = &LinkObject{
-			ID:   vl.Dataset.ID,
-			HRef: vl.Dataset.HRef,
-		}
-	}
-	if vl.Dimensions != nil {
-		dst.Dimensions = &LinkObject{
-			ID:   vl.Dimensions.ID,
-			HRef: vl.Dimensions.HRef,
-		}
-	}
-	if vl.Edition != nil {
-		dst.Edition = &LinkObject{
-			ID:   vl.Edition.ID,
-			HRef: vl.Edition.HRef,
-		}
-	}
-	if vl.Self != nil {
-		dst.Self = &LinkObject{
-			ID:   vl.Self.ID,
-			HRef: vl.Self.HRef,
-		}
-	}
-	if vl.Spatial != nil {
-		dst.Spatial = &LinkObject{
-			ID:   vl.Spatial.ID,
-			HRef: vl.Spatial.HRef,
-		}
-	}
-	if vl.Version != nil {
-		dst.Version = &LinkObject{
-			ID:   vl.Version.ID,
-			HRef: vl.Version.HRef,
-		}
-	}
-	return dst
 }
 
 // IsBasedOn refers to the Cantabular blob source
@@ -417,47 +205,6 @@ func CreateDataset(reader io.Reader) (*Dataset, error) {
 	}
 
 	return &dataset, nil
-}
-
-// CreateVersion manages the creation of a version from a reader
-func CreateVersion(reader io.Reader, datasetID string) (*Version, error) {
-	b, err := io.ReadAll(reader)
-	if err != nil {
-		return nil, errs.ErrUnableToReadMessage
-	}
-
-	log.Info(context.Background(), "DEBUG", log.Data{"body_create_version": string(b)})
-	var version Version
-	id, err := uuid.NewV4()
-	if err != nil {
-		return nil, err
-	}
-	version.ID = id.String()
-	version.DatasetID = datasetID
-
-	err = json.Unmarshal(b, &version)
-	if err != nil {
-		return nil, errs.ErrUnableToParseJSON
-	}
-
-	log.Info(context.Background(), "DEBUG", log.Data{"unmarshaled": version})
-
-	return &version, nil
-}
-
-// CreateDownloadList manages the creation of a list downloadable items from a reader
-func CreateDownloadList(reader io.Reader) (*DownloadList, error) {
-	b, err := io.ReadAll(reader)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to read response body")
-	}
-
-	var downloadList DownloadList
-	if err := json.Unmarshal(b, &downloadList); err != nil {
-		return nil, errors.Wrap(err, "failed to parse json to downloadList")
-	}
-
-	return &downloadList, nil
 }
 
 // CreateContact manages the creation of a contact from a reader
@@ -692,121 +439,4 @@ func ValidateDatasetType(ctx context.Context, datasetType string) (*DatasetType,
 		return nil, err
 	}
 	return &dataType, nil
-}
-
-// ValidateVersion checks the content of the version structure
-// TODO: Refactor this to reduce the complexity
-//
-//nolint:gocyclo,gocognit // high cyclomactic & cognitive complexity not in scope for maintenance
-func ValidateVersion(version *Version) error {
-	switch version.State {
-	case "":
-		return errs.ErrVersionMissingState
-	case EditionConfirmedState:
-	case PublishedState:
-		if version.CollectionID != "" {
-			return ErrPublishedVersionCollectionIDInvalid
-		}
-	case AssociatedState:
-		if version.CollectionID == "" {
-			return ErrAssociatedVersionCollectionIDInvalid
-		}
-	default:
-		return ErrVersionStateInvalid
-	}
-
-	var missingFields []string
-	var invalidFields []string
-
-	if version.ReleaseDate == "" {
-		missingFields = append(missingFields, "release_date")
-	}
-
-	if version.Downloads != nil {
-		if version.Downloads.XLS != nil {
-			if version.Downloads.XLS.HRef == "" {
-				missingFields = append(missingFields, "Downloads.XLS.HRef")
-			}
-			if version.Downloads.XLS.Size == "" {
-				missingFields = append(missingFields, "Downloads.XLS.Size")
-			}
-			if _, err := strconv.Atoi(version.Downloads.XLS.Size); err != nil {
-				invalidFields = append(invalidFields, "Downloads.XLS.Size not a number")
-			}
-		}
-
-		if version.Downloads.XLSX != nil {
-			if version.Downloads.XLSX.HRef == "" {
-				missingFields = append(missingFields, "Downloads.XLSX.HRef")
-			}
-			if version.Downloads.XLSX.Size == "" {
-				missingFields = append(missingFields, "Downloads.XLSX.Size")
-			}
-			if _, err := strconv.Atoi(version.Downloads.XLSX.Size); err != nil {
-				invalidFields = append(invalidFields, "Downloads.XLSX.Size not a number")
-			}
-		}
-
-		if version.Downloads.CSV != nil {
-			if version.Downloads.CSV.HRef == "" {
-				missingFields = append(missingFields, "Downloads.CSV.HRef")
-			}
-			if version.Downloads.CSV.Size == "" {
-				missingFields = append(missingFields, "Downloads.CSV.Size")
-			}
-			if _, err := strconv.Atoi(version.Downloads.CSV.Size); err != nil {
-				invalidFields = append(invalidFields, "Downloads.CSV.Size not a number")
-			}
-		}
-
-		if version.Downloads.CSVW != nil {
-			if version.Downloads.CSVW.HRef == "" {
-				missingFields = append(missingFields, "Downloads.CSVW.HRef")
-			}
-			if version.Downloads.CSVW.Size == "" {
-				missingFields = append(missingFields, "Downloads.CSVW.Size")
-			}
-			if _, err := strconv.Atoi(version.Downloads.CSVW.Size); err != nil {
-				invalidFields = append(invalidFields, "Downloads.CSVW.Size not a number")
-			}
-		}
-
-		if version.Downloads.TXT != nil {
-			if version.Downloads.TXT.HRef == "" {
-				missingFields = append(missingFields, "Downloads.TXT.HRef")
-			}
-			if version.Downloads.TXT.Size == "" {
-				missingFields = append(missingFields, "Downloads.TXT.Size")
-			}
-			if _, err := strconv.Atoi(version.Downloads.TXT.Size); err != nil {
-				invalidFields = append(invalidFields, "Downloads.TXT.Size not a number")
-			}
-		}
-	}
-
-	if missingFields != nil {
-		return fmt.Errorf("missing mandatory fields: %v", missingFields)
-	}
-
-	if invalidFields != nil {
-		return fmt.Errorf("invalid fields: %v", invalidFields)
-	}
-
-	return nil
-}
-
-// ParseAndValidateVersionNumber checks the version is a positive integer above 0
-func ParseAndValidateVersionNumber(ctx context.Context, version string) (int, error) {
-	versionNumber, err := strconv.Atoi(version)
-	if err != nil {
-		log.Error(ctx, "invalid version provided", err, log.Data{"version": version})
-		return versionNumber, errs.ErrInvalidVersion
-	}
-
-	if !(versionNumber > 0) {
-		log.Error(ctx, "version is not a positive integer", errs.ErrInvalidVersion, log.Data{"version": version})
-		return versionNumber, errs.ErrInvalidVersion
-	}
-
-	return versionNumber, nil
 }
